@@ -18,57 +18,54 @@
     "test:watch": "npm test -- --watch"
   },
   "dependencies": {
-    "bootstrap": "~3.3.0",
-    "chroma-js": "^1.2.2",
     "ansi-style-parser": "^1.0.1",
     "babel-runtime": "^6.9.2",
+    "bootstrap": "~3.3.0",
     "bootstrap-sass": "^3.3.6",
+    "chroma-js": "^1.2.2",
     "classnames": "^2.2.3",
     "clipboard": "^1.5.8",
-    "core-js": "^1.2.7",
-    "eonasdan-bootstrap-datetimepicker": "^4.15.35",
-    "fuse.js": "~1.2.2",
+    "core-js": "^2.4.1",
+    "fuse.js": "^3.0.5",
     "fuzzy": "git://github.com/HubSpot/fuzzy.git#v0.1.1-plus-return-match-structure",
     "history": "^2.1.2",
     "humanize-plus": "^1.8.1",
     "isomorphic-fetch": "^2.2.1",
-    "jquery": "~1.11.1",
+    "jquery": "^3.2.1",
     "juration": "*",
     "keymaster": "^1.6.2",
     "messenger": "git://github.com/HubSpot/messenger.git#set-main-field",
     "micromatch": "^2.3.8",
     "moment": "^2.13.0",
     "parseurl": "^1.3.1",
-    "promise-polyfill": "^5.2.0",
+    "promise-polyfill": "^6.0.2",
+    "prop-types": "^15.5.10",
     "q": "^1.4.1",
     "raven-js": "^3.9.1",
-    "react": "^15.1.0",
-    "react-bootstrap": "^0.30.0",
+    "react": "^15.6.1",
+    "react-bootstrap": "^0.31.0",
     "react-bootstrap-datetimepicker": "0.0.22",
-    "react-dom": "^15.1.0",
-    "react-interval": "^1.2.1",
-    "react-json-tree": "^0.8.0",
-    "react-list": "~0.7.3",
-    "react-redux": "^4.4.5",
+    "react-circular-progressbar": "^0.1.3",
+    "react-copy-to-clipboard": "^5.0.0",
+    "react-dom": "^15.6.1",
+    "react-interval": "^2.0.1",
+    "react-json-tree": "^0.10.9",
+    "react-redux": "^5.0.5",
     "react-router": "^2.5.2",
     "react-router-redux": "^4.0.5",
-    "react-select": "1.0.0-beta14",
+    "react-select": "^1.0.0-rc.5",
     "react-tagsinput": "^3.13.0",
     "react-typeahead": "git://github.com/HubSpot/react-typeahead.git#hubspot-3",
     "react-waypoint": "^2.0.3",
-    "react-copy-to-clipboard": "^4.2.3",
-    "react-circular-progressbar": "^0.1.3",
     "redux": "^3.5.2",
     "redux-form": "^5.3.0",
-    "redux-logger": "^2.6.1",
     "redux-thunk": "^2.0.1",
-    "reselect": "^2.5.1",
-    "select2": "~3.5.1",
+    "reselect": "^3.0.1",
+    "select2": "^4.0.3",
     "singularityui-tailer": "0.3.11",
-    "typeahead.js": "~0.10.4",
+    "typeahead.js": "^0.11.1",
     "underscore": "~1.8.0",
-    "uuid": "^3.0.0",
-    "vex-js": "git://github.com/HubSpot/vex.git#v2.3.0-plus-beforeClose-hook"
+    "uuid": "^3.0.0"
   },
   "license": "Apache-2.0",
   "devDependencies": {
@@ -102,7 +99,7 @@
     "gulp-util": "^3.0.7",
     "mocha": "^3.1.2",
     "nib": "^1.1.0",
-    "node-sass": "^3.8.0",
+    "node-sass": "^4.5.3",
     "react-copy-to-clipboard": "^4.2.3",
     "react-circular-progressbar": "^0.1.3",
     "react-hot-loader": "^3.0.0-beta.2",
