@@ -4,6 +4,7 @@ import Tooltip from 'react-bootstrap/lib/Tooltip';
 import OverlayTrigger from 'react-bootstrap/lib/OverlayTrigger';
 
 const Breakdown = ({total, data}) => {
+
   const sections = data.map((item, key) => {
     return (
       <Link key={key} to={item.link}>
@@ -12,7 +13,7 @@ const Breakdown = ({total, data}) => {
             data-type="column"
             data-state-attribute={item.attribute}
             style={{height: `${item.percent}%`}}
-            className={`chart__data-point chart-fill-${item.type}`}
+            className={`chart__data-point bg-${item.type}`}
             data-original-title={`${item.count} ${item.label}`}
           />
         </OverlayTrigger>
@@ -33,10 +34,11 @@ const Breakdown = ({total, data}) => {
 Breakdown.propTypes = {
   total: React.PropTypes.number.isRequired,
   data: React.PropTypes.arrayOf(React.PropTypes.shape({
+    attribute: React.PropTypes.string.isRequired,
     count: React.PropTypes.number.isRequired,
     type: React.PropTypes.string.isRequired,
     label: React.PropTypes.string.isRequired,
-    link: React.PropTypes.string.isRequired,
+    link: React.PropTypes.string,
     percent: React.PropTypes.number.isRequired
   })).isRequired
 };
