@@ -5,19 +5,19 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class SingularityTaskCurrentUsage {
 
-  private final long memoryRssBytes;
+  private final long memoryTotalBytes;
   private final long timestamp;
   private final double cpusUsed;
 
   @JsonCreator
-  public SingularityTaskCurrentUsage(@JsonProperty("memoryRssBytes") long memoryRssBytes, @JsonProperty("long") long timestamp, @JsonProperty("cpusUsed") double cpusUsed) {
-    this.memoryRssBytes = memoryRssBytes;
+  public SingularityTaskCurrentUsage(@JsonProperty("memoryTotalBytes") long memoryTotalBytes, @JsonProperty("long") long timestamp, @JsonProperty("cpusUsed") double cpusUsed) {
+    this.memoryTotalBytes = memoryTotalBytes;
     this.timestamp = timestamp;
     this.cpusUsed = cpusUsed;
   }
 
-  public long getMemoryRssBytes() {
-    return memoryRssBytes;
+  public long getMemoryTotalBytes() {
+    return memoryTotalBytes;
   }
 
   public long getTimestamp() {
@@ -30,7 +30,7 @@ public class SingularityTaskCurrentUsage {
 
   @Override
   public String toString() {
-    return "SingularityTaskCurrentUsage [memoryRssBytes=" + memoryRssBytes + ", timestamp=" + timestamp + ", cpusUsed=" + cpusUsed + "]";
+    return "SingularityTaskCurrentUsage [memoryTotalBytes=" + memoryTotalBytes + ", timestamp=" + timestamp + ", cpusUsed=" + cpusUsed + "]";
   }
 
 }
