@@ -6,14 +6,14 @@ import java.io.IOException;
 import com.github.jknack.handlebars.Helper;
 import com.github.jknack.handlebars.Options;
 
-public class IfHasNewLinesHelper implements Helper<Object> {
+public class IfHasNewLinesOrBackticksHelper implements Helper<Object> {
 
-  public static final String NAME = "ifHasNewLines";
+  public static final String NAME = "ifHasNewLinesOrBackticks";
 
   @SuppressWarnings("unchecked")
   @Override
   public CharSequence apply(Object context, Options options) throws IOException {
-    if (context.toString().contains("\n")) {
+    if (context.toString().contains("\n") || context.toString().contains("`")) {
       return options.fn();
     } else {
       return options.inverse();
