@@ -25,11 +25,9 @@ import org.slf4j.LoggerFactory;
 
 import com.codahale.metrics.annotation.Timed;
 import com.google.common.base.Optional;
-import com.google.common.base.Predicates;
 import com.google.common.base.Throwables;
 import com.google.common.collect.HashMultiset;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.Iterables;
 import com.google.common.collect.ListMultimap;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
@@ -49,6 +47,7 @@ import com.hubspot.singularity.SingularityDeployMarker;
 import com.hubspot.singularity.SingularityDeployProgress;
 import com.hubspot.singularity.SingularityDeployStatistics;
 import com.hubspot.singularity.SingularityDeployStatisticsBuilder;
+import com.hubspot.singularity.SingularityKilledTaskIdRecord;
 import com.hubspot.singularity.SingularityMachineAbstraction;
 import com.hubspot.singularity.SingularityPendingDeploy;
 import com.hubspot.singularity.SingularityPendingRequest;
@@ -96,9 +95,11 @@ public class SingularityScheduler {
 
   private final SingularityMailer mailer;
 
+  private final SingularityLeaderCache leaderCache;
+
   @Inject
   public SingularityScheduler(TaskRequestManager taskRequestManager, SingularityConfiguration configuration, SingularityCooldown cooldown, DeployManager deployManager,
-    TaskManager taskManager, RequestManager requestManager, SlaveManager slaveManager, RackManager rackManager, SingularityMailer mailer) {
+    TaskManager taskManager, RequestManager requestManager, SlaveManager slaveManager, RackManager rackManager, SingularityMailer mailer, SingularityLeaderCache leaderCache) {
     this.taskRequestManager = taskRequestManager;
     this.configuration = configuration;
     this.deployManager = deployManager;
@@ -108,6 +109,7 @@ public class SingularityScheduler {
     this.rackManager = rackManager;
     this.mailer = mailer;
     this.cooldown = cooldown;
+    this.leaderCache = leaderCache;
   }
 
   private void cleanupTaskDueToDecomission(final Map<String, Optional<String>> requestIdsToUserToReschedule, final Set<SingularityTaskId> matchingTaskIds, SingularityTask task,
@@ -131,13 +133,13 @@ public class SingularityScheduler {
   }
 
   @Timed
-  public void checkForDecomissions(SingularitySchedulerStateCache stateCache) {
+  public void checkForDecomissions() {
     final long start = System.currentTimeMillis();
 
     final Map<String, Optional<String>> requestIdsToUserToReschedule = Maps.newHashMap();
     final Set<SingularityTaskId> matchingTaskIds = Sets.newHashSet();
 
-    final Collection<SingularityTaskId> activeTaskIds = stateCache.getActiveTaskIds();
+    final Collection<SingularityTaskId> activeTaskIds = leaderCache.getActiveTaskIds();
 
     final Map<SingularitySlave, MachineState> slaves = getDefaultMap(slaveManager.getObjectsFiltered(MachineState.STARTING_DECOMMISSION));
 
@@ -215,7 +217,7 @@ public class SingularityScheduler {
   }
 
   @Timed
-  public void drainPendingQueue(final SingularitySchedulerStateCache stateCache) {
+  public void drainPendingQueue() {
     final long start = System.currentTimeMillis();
     final ImmutableList<SingularityPendingRequest> pendingRequests = ImmutableList.copyOf(requestManager.getPendingRequests());
 
@@ -250,9 +252,11 @@ public class SingularityScheduler {
 
       Optional<SingularityRequestDeployState> maybeRequestDeployState = deployManager.getRequestDeployState(requestId);
       Optional<SingularityPendingDeploy> maybePendingDeploy = deployManager.getPendingDeploy(requestId);
-      List<SingularityTaskId> matchingTaskIds = getMatchingTaskIds(stateCache, maybeRequest.get().getRequest(), deployKey);
+      List<SingularityTaskId> matchingTaskIds = getMatchingTaskIds(maybeRequest.get().getRequest(), deployKey);
 
       List<SingularityPendingRequest> effectivePendingRequests = new ArrayList<>();
+
+      // Things that are closest to now (ie smaller timestamps) should come first in the queue
       pendingRequestsForDeploy.sort(Comparator.comparingLong(SingularityPendingRequest::getTimestamp));
       int scheduledTasks = 0;
       for (SingularityPendingRequest pendingRequest : pendingRequestsForDeploy) {
@@ -275,20 +279,26 @@ public class SingularityScheduler {
         if (effectivePendingRequests.isEmpty()) {
           effectivePendingRequests.add(pendingRequest);
           RequestState requestState = checkCooldown(maybeRequest.get().getState(), maybeRequest.get().getRequest(), deployStatistics);
-          scheduledTasks += scheduleTasks(stateCache, maybeRequest.get().getRequest(), requestState,
+          scheduledTasks += scheduleTasks(maybeRequest.get().getRequest(), requestState,
               deployStatistics, pendingRequest, matchingTaskIds, maybePendingDeploy);
           requestManager.deletePendingRequest(pendingRequest);
         } else if (pendingRequest.getPendingType() == PendingType.IMMEDIATE) {
           effectivePendingRequests.add(pendingRequest);
           RequestState requestState = checkCooldown(maybeRequest.get().getState(), maybeRequest.get().getRequest(), deployStatistics);
-          scheduledTasks += scheduleTasks(stateCache, maybeRequest.get().getRequest(), requestState,
+          scheduledTasks += scheduleTasks(maybeRequest.get().getRequest(), requestState,
               deployStatistics, pendingRequest, matchingTaskIds, maybePendingDeploy);
           requestManager.deletePendingRequest(pendingRequest);
         } else if (pendingRequest.getPendingType() == PendingType.ONEOFF) {
           effectivePendingRequests.add(pendingRequest);
           RequestState requestState = checkCooldown(maybeRequest.get().getState(), maybeRequest.get().getRequest(), deployStatistics);
-          scheduledTasks += scheduleTasks(stateCache, maybeRequest.get().getRequest(), requestState,
+          scheduledTasks += scheduleTasks(maybeRequest.get().getRequest(), requestState,
               deployStatistics, pendingRequest, matchingTaskIds, maybePendingDeploy);
+          requestManager.deletePendingRequest(pendingRequest);
+        } else if (updatedRequest.isScheduled()
+            && (pendingRequest.getPendingType() == PendingType.NEW_DEPLOY
+            || pendingRequest.getPendingType() == PendingType.TASK_DONE)) {
+          // If we are here, there is already an immediate of run of the scheduled task launched. Drop anything that would
+          // leave a second instance of the request in the pending queue.
           requestManager.deletePendingRequest(pendingRequest);
         } else {
           // Any other subsequent requests are not honored until after the pending queue is cleared.
@@ -392,39 +402,48 @@ public class SingularityScheduler {
   }
 
   private void deleteScheduledTasks(final Collection<SingularityPendingTask> scheduledTasks, SingularityPendingRequest pendingRequest) {
-    for (SingularityPendingTask task : Iterables
-      .filter(scheduledTasks, Predicates.and(SingularityPendingTask.matchingRequest(pendingRequest.getRequestId()), SingularityPendingTask.matchingDeploy(pendingRequest.getDeployId())))) {
+    List<SingularityPendingTask> tasksForDeploy = scheduledTasks
+        .stream()
+        .filter(task -> pendingRequest.getRequestId().equals(task.getPendingTaskId().getRequestId()))
+        .filter(task -> pendingRequest.getDeployId().equals(task.getPendingTaskId().getDeployId()))
+        .collect(Collectors.toList());
+
+    for (SingularityPendingTask task : tasksForDeploy) {
       LOG.debug("Deleting pending task {} in order to reschedule {}", task.getPendingTaskId().getId(), pendingRequest);
       taskManager.deletePendingTask(task.getPendingTaskId());
     }
   }
 
-  private List<SingularityTaskId> getMatchingTaskIds(SingularitySchedulerStateCache stateCache, SingularityRequest request, SingularityDeployKey deployKey) {
+  private List<SingularityTaskId> getMatchingTaskIds(SingularityRequest request, SingularityDeployKey deployKey) {
     if (request.isLongRunning()) {
+      Set<SingularityTaskId> killedTaskIds = leaderCache.getKilledTasks().stream()
+          .map(SingularityKilledTaskIdRecord::getTaskId)
+          .collect(Collectors.toSet());
+
       List<SingularityTaskId> matchingTaskIds = new ArrayList<>();
-      for (SingularityTaskId taskId : stateCache.getActiveTaskIdsForRequest(deployKey.getRequestId())) {
+      for (SingularityTaskId taskId : leaderCache.getActiveTaskIdsForRequest(deployKey.getRequestId())) {
         if (!taskId.getDeployId().equals(deployKey.getDeployId())) {
           continue;
         }
-        if (stateCache.getCleaningTasks().contains(taskId)) {
+        if (leaderCache.getCleanupTaskIds().contains(taskId)) {
           continue;
         }
-        if (stateCache.getKilledTasks().contains(taskId)) {
+        if (killedTaskIds.contains(taskId)) {
           continue;
         }
         matchingTaskIds.add(taskId);
       }
       return matchingTaskIds;
     } else {
-      return new ArrayList<>(stateCache.getActiveTaskIdsForRequest(deployKey.getRequestId()));
+      return new ArrayList<>(leaderCache.getActiveTaskIdsForRequest(deployKey.getRequestId()));
     }
   }
 
-  private int scheduleTasks(SingularitySchedulerStateCache stateCache, SingularityRequest request, RequestState state,
+  private int scheduleTasks(SingularityRequest request, RequestState state,
                             SingularityDeployStatistics deployStatistics, SingularityPendingRequest pendingRequest,
                             List<SingularityTaskId> matchingTaskIds, Optional<SingularityPendingDeploy> maybePendingDeploy) {
     if (request.getRequestType() != RequestType.ON_DEMAND) {
-      deleteScheduledTasks(stateCache.getScheduledTasks(), pendingRequest);
+      deleteScheduledTasks(leaderCache.getPendingTasks(), pendingRequest);
     }
 
     final int numMissingInstances = getNumMissingInstances(matchingTaskIds, request, pendingRequest, maybePendingDeploy);
@@ -453,7 +472,7 @@ public class SingularityScheduler {
 
       if (request.isRackSensitive() && configuration.isRebalanceRacksOnScaleDown()) {
         List<SingularityTaskId> extraCleanedTasks = new ArrayList<>();
-        int numActiveRacks = stateCache.getNumActiveRacks();
+        int numActiveRacks = rackManager.getNumActive();
         double perRack = request.getInstancesSafe() / (double) numActiveRacks;
 
         Multiset<String> countPerRack = HashMultiset.create();
@@ -515,7 +534,7 @@ public class SingularityScheduler {
   }
 
   private Optional<PendingType> handleCompletedTaskWithStatistics(Optional<SingularityTask> task, SingularityTaskId taskId, long timestamp, ExtendedTaskState state,
-    SingularityDeployStatistics deployStatistics, SingularityCreateResult taskHistoryUpdateCreateResult, SingularitySchedulerStateCache stateCache, Protos.TaskStatus status) {
+    SingularityDeployStatistics deployStatistics, SingularityCreateResult taskHistoryUpdateCreateResult, Protos.TaskStatus status) {
     final Optional<SingularityRequestWithState> maybeRequestWithState = requestManager.getRequest(taskId.getRequestId());
     final Optional<SingularityPendingDeploy> maybePendingDeploy = deployManager.getPendingDeploy(taskId.getRequestId());
 
@@ -578,7 +597,7 @@ public class SingularityScheduler {
       Optional.<String>absent());
 
     SingularityDeployKey deployKey = new SingularityDeployKey(taskId.getRequestId(), taskId.getDeployId());
-    scheduleTasks(stateCache, request, requestState, deployStatistics, pendingRequest, getMatchingTaskIds(stateCache, request, deployKey), maybePendingDeploy);
+    scheduleTasks(request, requestState, deployStatistics, pendingRequest, getMatchingTaskIds(request, deployKey), maybePendingDeploy);
 
     return Optional.of(pendingType);
   }
@@ -595,28 +614,27 @@ public class SingularityScheduler {
 
   @Timed
   public void handleCompletedTask(Optional<SingularityTask> task, SingularityTaskId taskId, boolean wasActive, long timestamp, ExtendedTaskState state,
-    SingularityCreateResult taskHistoryUpdateCreateResult, SingularitySchedulerStateCache stateCache, Protos.TaskStatus status) {
+    SingularityCreateResult taskHistoryUpdateCreateResult, Protos.TaskStatus status) {
     final SingularityDeployStatistics deployStatistics = getDeployStatistics(taskId.getRequestId(), taskId.getDeployId());
 
     if (wasActive) {
       taskManager.deleteActiveTask(taskId.getId());
-      stateCache.getActiveTaskIds().remove(taskId);
     }
 
     if (!task.isPresent() || task.get().getTaskRequest().getRequest().isLoadBalanced()) {
       taskManager.createLBCleanupTask(taskId);
     }
 
-    final Optional<PendingType> scheduleResult = handleCompletedTaskWithStatistics(task, taskId, timestamp, state, deployStatistics, taskHistoryUpdateCreateResult, stateCache, status);
+    final Optional<PendingType> scheduleResult = handleCompletedTaskWithStatistics(task, taskId, timestamp, state, deployStatistics, taskHistoryUpdateCreateResult, status);
 
     if (taskHistoryUpdateCreateResult == SingularityCreateResult.EXISTED) {
       return;
     }
 
-    updateDeployStatistics(deployStatistics, taskId, timestamp, state, scheduleResult);
+    updateDeployStatistics(deployStatistics, taskId, task, timestamp, state, scheduleResult);
   }
 
-  private void updateDeployStatistics(SingularityDeployStatistics deployStatistics, SingularityTaskId taskId, long timestamp, ExtendedTaskState state, Optional<PendingType> scheduleResult) {
+  private void updateDeployStatistics(SingularityDeployStatistics deployStatistics, SingularityTaskId taskId, Optional<SingularityTask> task, long timestamp, ExtendedTaskState state, Optional<PendingType> scheduleResult) {
     SingularityDeployStatisticsBuilder bldr = deployStatistics.toBuilder();
 
     if (!state.isFailed()) {
@@ -627,6 +645,22 @@ public class SingularityScheduler {
       } else {
         bldr.setAverageRuntimeMillis(Optional.of(timestamp - taskId.getStartedAt()));
       }
+    }
+
+    if (task.isPresent()) {
+      long dueTime = task.get().getTaskRequest().getPendingTask().getPendingTaskId().getNextRunAt();
+      long startedAt = taskId.getStartedAt();
+
+      if (bldr.getAverageSchedulingDelayMillis().isPresent()) {
+        long newAverageSchedulingDelayMillis = (bldr.getAverageSchedulingDelayMillis().get() * bldr.getNumTasks() + (startedAt - dueTime)) / (bldr.getNumTasks() + 1);
+        bldr.setAverageSchedulingDelayMillis(Optional.of(newAverageSchedulingDelayMillis));
+      } else {
+        bldr.setAverageSchedulingDelayMillis(Optional.of(startedAt - dueTime));
+      }
+
+      final SingularityDeployStatistics newStatistics = bldr.build();
+
+      deployManager.saveDeployStatistics(newStatistics);
     }
 
     bldr.setNumTasks(bldr.getNumTasks() + 1);
