@@ -17,6 +17,7 @@ import com.hubspot.mesos.Resources;
 import com.hubspot.singularity.DeployState;
 import com.hubspot.singularity.LoadBalancerRequestType;
 import com.hubspot.singularity.RequestState;
+import com.hubspot.singularity.RequestType;
 import com.hubspot.singularity.SingularityDeploy;
 import com.hubspot.singularity.SingularityDeployBuilder;
 import com.hubspot.singularity.SingularityDeployProgress;
@@ -30,6 +31,7 @@ import com.hubspot.singularity.SingularityTaskId;
 import com.hubspot.singularity.SingularityUpdatePendingDeployRequest;
 import com.hubspot.singularity.TaskCleanupType;
 import com.hubspot.singularity.api.SingularityDeployRequest;
+import com.hubspot.singularity.api.SingularityRunNowRequest;
 
 public class SingularityDeploysTest extends SingularitySchedulerTestBase {
 
@@ -119,7 +121,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
 
     deploy(secondDeployId);
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     Assert.assertEquals(2, taskManager.getActiveTaskIds().size());
     Assert.assertEquals(2, taskManager.getPendingTaskIds().size());
@@ -153,7 +155,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
 
     deploy(secondDeployId, Optional.<Boolean> absent(), Optional.of(1), Optional.<Boolean> absent(), false);
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
 
     resourceOffers();
@@ -178,7 +180,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
     Assert.assertFalse(deployProgressStepTwo.isStepComplete());
     Assert.assertEquals(2, deployProgressStepTwo.getTargetActiveInstances());
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
 
     resourceOffers();
@@ -206,7 +208,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
     launchTask(request, firstDeploy, 2, TaskState.TASK_RUNNING);
     deploy(secondDeployId, Optional.<Boolean> absent(), Optional.of(1), Optional.<Boolean> absent(), false);
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     resourceOffers();
     SingularityTaskId firstNewTaskId = taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).get(0);
     statusUpdate(taskManager.getTask(firstNewTaskId).get(), TaskState.TASK_RUNNING);
@@ -215,7 +217,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
     cleaner.drainCleanupQueue();
     statusUpdate(firstTask, TaskState.TASK_KILLED);
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     resourceOffers();
     // End in-progress incremental deploy setup
 
@@ -248,7 +250,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
 
     deploy(secondDeployId);
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     resourceOffers();
 
     Assert.assertEquals(2, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
@@ -258,7 +260,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
     }
 
     requestResource.postRequest(request.toBuilder().setInstances(Optional.of(1)).build());
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     Assert.assertEquals(1, taskManager.getCleanupTaskIds().size());
 
@@ -266,7 +268,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
     Assert.assertEquals(2, taskManager.getCleanupTaskIds().size());
 
     // Extra task from the new deploy should get cleaned up as well
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     Assert.assertEquals(3, taskManager.getCleanupTaskIds().size());
   }
 
@@ -285,7 +287,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
 
     deploy(secondDeployId, Optional.<Boolean>absent(), Optional.of(1), Optional.of(false), false);
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
 
     resourceOffers();
@@ -320,7 +322,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
     Assert.assertFalse(deployProgressStepTwo.isStepComplete());
     Assert.assertEquals(2, deployProgressStepTwo.getTargetActiveInstances());
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
 
     resourceOffers();
@@ -352,7 +354,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
 
     deploy(secondDeployId, Optional.<Boolean>absent(), Optional.of(2), Optional.<Boolean> absent(), false);
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     Assert.assertEquals(2, taskManager.getPendingTaskIds().size());
 
     resourceOffers();
@@ -380,7 +382,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
     Assert.assertFalse(deployProgressStepTwo.isStepComplete());
     Assert.assertEquals(4, deployProgressStepTwo.getTargetActiveInstances());
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     Assert.assertEquals(2, taskManager.getPendingTaskIds().size());
 
     resourceOffers();
@@ -407,7 +409,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
 
     deploy(secondDeployId, Optional.<Boolean>absent(), Optional.of(1), Optional.of(false), false);
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
 
     resourceOffers();
@@ -437,7 +439,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
 
     deploy(secondDeployId, Optional.<Boolean>absent(), Optional.of(1), Optional.of(false), false);
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
 
     resourceOffers();
@@ -523,7 +525,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
 
     deploy(secondDeployId, Optional.<Boolean>absent(), Optional.of(1), Optional.<Boolean> absent(), true);
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
 
     resourceOffers();
@@ -560,7 +562,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
     Assert.assertFalse(pendingDeploy.getDeployProgress().get().isStepComplete());
     Assert.assertEquals(2, pendingDeploy.getDeployProgress().get().getTargetActiveInstances());
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
 
     resourceOffers();
@@ -603,7 +605,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
 
     deploy(secondDeployId, Optional.<Boolean> absent(), Optional.of(1), Optional.<Boolean> absent(), false);
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
 
     resourceOffers();
@@ -626,7 +628,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
     deployResource.cancelDeploy(requestId, secondDeployId);
     deployChecker.checkDeploys();
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     List<SingularityPendingTaskId> pendingTaskIds = taskManager.getPendingTaskIds();
     Assert.assertEquals(1, pendingTaskIds.size());
     Assert.assertEquals(firstDeployId, pendingTaskIds.get(0).getDeployId());
@@ -661,7 +663,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
 
     deploy("nextDeployId");
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     // no second task should be scheduled
 
@@ -671,7 +673,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
     Assert.assertTrue(!taskManager.getCleanupTaskIds().isEmpty());
 
     statusUpdate(firstTask, TaskState.TASK_FINISHED);
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     cleaner.drainCleanupQueue();
 
     Assert.assertTrue(!taskManager.getPendingTasks().isEmpty());
@@ -694,11 +696,11 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
     Assert.assertTrue(!taskManager.getPendingTaskIds().isEmpty());
 
     deploy("d2");
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     deployChecker.checkDeploys();
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     Assert.assertTrue(!taskManager.getPendingTaskIds().isEmpty());
   }
@@ -724,7 +726,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
 
     Assert.assertTrue(requestManager.getRequest(requestId).get().getState() == RequestState.DEPLOYING_TO_UNPAUSE);
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1")));
     statusUpdate(taskManager.getActiveTasks().get(0), TaskState.TASK_FAILED);
 
@@ -740,7 +742,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
 
     Assert.assertTrue(requestManager.getRequest(requestId).get().getState() == RequestState.DEPLOYING_TO_UNPAUSE);
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1")));
 
     statusUpdate(taskManager.getActiveTasks().get(0), TaskState.TASK_RUNNING);
@@ -781,7 +783,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
 
     saveAndSchedule(request.toBuilder().setInstances(Optional.of(2)));
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     Assert.assertEquals(2, taskManager.getPendingTaskIds().size());
     Assert.assertEquals(2, requestManager.getRequest(requestId).get().getRequest().getInstancesSafe());
@@ -795,7 +797,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
     deployResource.deploy(new SingularityDeployRequest(deploy, Optional.<Boolean>absent(), Optional.<String>absent(), Optional.of(newRequest)));
 
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     List<SingularityPendingTaskId> pendingTaskIdsForNewDeploy = new ArrayList<>();
     for (SingularityPendingTaskId pendingTaskId : taskManager.getPendingTaskIds()) {
@@ -862,7 +864,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
 
     deploy(secondDeployId, Optional.<Boolean>absent(), Optional.of(1), Optional.<Boolean>absent(), false);
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     resourceOffers();
     SingularityTaskId firstNewTaskId = taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).get(0);
     statusUpdate(taskManager.getTask(firstNewTaskId).get(), TaskState.TASK_RUNNING);
@@ -876,7 +878,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
     statusUpdate(firstTask, TaskState.TASK_KILLED);
 
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     resourceOffers();
     for (SingularityTaskId taskId : taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId)) {
       statusUpdate(taskManager.getTask(taskId).get(), TaskState.TASK_RUNNING);
@@ -891,7 +893,7 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
     statusUpdate(secondTask, TaskState.TASK_KILLED);
 
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     resourceOffers();
     for (SingularityTaskId taskId : taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId)) {
       statusUpdate(taskManager.getTask(taskId).get(), TaskState.TASK_RUNNING);
@@ -901,5 +903,59 @@ public class SingularityDeploysTest extends SingularitySchedulerTestBase {
     deployProgress = deployManager.getPendingDeploy(requestId).get().getDeployProgress().get();
     Assert.assertEquals(3, deployProgress.getTargetActiveInstances());
     Assert.assertEquals(3, deployProgress.getCurrentActiveInstances());
+  }
+
+  @Test
+  public void testDeployWithImmediateRunIsLaunchedImmediately() {
+    initRequestWithType(RequestType.SCHEDULED, false);
+    String deployId = "d1";
+
+    SingularityRunNowRequest runNowRequest = new SingularityRunNowRequest(Optional.of("Message"), Optional.absent(), Optional.absent(), Optional.absent(), Optional.absent());
+    SingularityDeploy deploy = new SingularityDeployBuilder(requestId, deployId)
+        .setRunImmediately(Optional.of(runNowRequest))
+        .setCommand(Optional.of("printenv > tmp.txt"))
+        .build();
+    SingularityDeployRequest deployRequest = new SingularityDeployRequest(deploy, Optional.absent(), Optional.absent());
+    deployResource.deploy(deployRequest);
+    deployChecker.checkDeploys();
+
+    scheduler.drainPendingQueue();
+    resourceOffers();
+
+    Assert.assertEquals(1, taskManager.getNumActiveTasks());
+    Assert.assertEquals(0, taskManager.getNumScheduledTasks());
+    SingularityTaskId taskId = taskManager.getActiveTaskIdsForDeploy(requestId, deployId).get(0);
+    SingularityTask task = taskManager.getTask(taskId).get();
+
+    Assert.assertEquals("printenv > tmp.txt", task.getMesosTask().getCommand().getValue());
+  }
+
+  @Test
+  public void testDeployWithImmediateRunSchedulesAfterRunningImmediately() {
+    initRequestWithType(RequestType.SCHEDULED, false);
+    String deployId = "d1";
+
+    SingularityRunNowRequest runNowRequest = new SingularityRunNowRequest(Optional.of("Message"), Optional.absent(), Optional.absent(), Optional.absent(), Optional.absent());
+    SingularityDeploy deploy = new SingularityDeployBuilder(requestId, deployId)
+        .setRunImmediately(Optional.of(runNowRequest))
+        .setCommand(Optional.of("printenv > tmp.txt"))
+        .build();
+    SingularityDeployRequest deployRequest = new SingularityDeployRequest(deploy, Optional.absent(), Optional.absent());
+    deployResource.deploy(deployRequest);
+    deployChecker.checkDeploys();
+
+    scheduler.drainPendingQueue();
+    resourceOffers();
+
+    SingularityTaskId taskId = taskManager.getActiveTaskIdsForDeploy(requestId, deployId).get(0);
+    SingularityTask task = taskManager.getTask(taskId).get();
+    statusUpdate(task, TaskState.TASK_RUNNING);
+    statusUpdate(task, TaskState.TASK_FINISHED);
+
+    scheduler.drainPendingQueue();
+    resourceOffers();
+
+    Assert.assertEquals(0, taskManager.getNumActiveTasks());
+    Assert.assertEquals(1, taskManager.getNumScheduledTasks());
   }
 }
