@@ -1,5 +1,7 @@
 package com.hubspot.singularity.scheduler;
 
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -16,6 +18,7 @@ import org.apache.mesos.Protos.SlaveID;
 import org.apache.mesos.Protos.TaskID;
 import org.apache.mesos.Protos.TaskState;
 import org.apache.mesos.Protos.TaskStatus;
+import org.assertj.core.api.Assertions;
 import org.junit.Assert;
 import org.junit.Test;
 import org.mockito.Matchers;
@@ -87,6 +90,7 @@ import com.hubspot.singularity.mesos.OfferCache;
 import com.hubspot.singularity.mesos.SingularityMesosTaskPrioritizer;
 import com.hubspot.singularity.scheduler.SingularityDeployHealthHelper.DeployHealth;
 import com.hubspot.singularity.scheduler.SingularityTaskReconciliation.ReconciliationState;
+import com.jayway.awaitility.Awaitility;
 
 public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
   @Inject
@@ -205,7 +209,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     initRequest();
 
     firstDeploy = initAndFinishDeploy(request, new SingularityDeployBuilder(request.getId(), firstDeployId)
-        .setCommand(Optional.of("sleep 100")).setResources(Optional.of(new Resources(1, 128, 2, 0)))
+        .setCommand(Optional.of("sleep 100")), Optional.of(new Resources(1, 128, 2, 0))
     );
 
     requestManager.addToPendingQueue(
@@ -271,7 +275,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     requestManager.addToPendingQueue(new SingularityPendingRequest(requestId, secondDeployId, System.currentTimeMillis(), Optional.<String> absent(), PendingType.NEW_DEPLOY,
         Optional.<Boolean> absent(), Optional.<String> absent()));
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     // we expect there to be 3 pending tasks :
 
@@ -487,7 +491,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     initFirstDeploy();
 
     requestResource.postRequest(request.toBuilder().setInstances(Optional.of(3)).build());
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     List<Offer> oneOffer = Arrays.asList(createOffer(12, 1024));
     sms.resourceOffers(driver, oneOffer);
@@ -503,7 +507,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     initFirstDeploy();
 
     requestResource.postRequest(request.toBuilder().setInstances(Optional.of(10)).build());
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     sms.resourceOffers(driver, Arrays.asList(createOffer(2, 1024), createOffer(1, 1024)));
 
@@ -517,11 +521,11 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     initFirstDeploy();
 
     requestResource.postRequest(request.toBuilder().setInstances(Optional.of(15)).build());
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     sms.resourceOffers(driver, Arrays.asList(createOffer(20, 1024), createOffer(20, 1024)));
 
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 15);
+    Assert.assertEquals(15, taskManager.getActiveTaskIds().size());
 
     Set<String> offerIds = Sets.newHashSet();
 
@@ -529,7 +533,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
       offerIds.addAll(activeTask.getOffers().stream().map((o) -> o.getId().getValue()).collect(Collectors.toList()));
     }
 
-    Assert.assertTrue(offerIds.size() == 2);
+    Assert.assertEquals(2, offerIds.size());
   }
 
   @Test
@@ -541,7 +545,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     // cause it to be pending
     requestResource.postRequest(request.toBuilder().setQuartzSchedule(Optional.of(schedule)).build());
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     Assert.assertTrue(requestResource.getActiveRequests(false).isEmpty());
     Assert.assertTrue(requestManager.getRequest(requestId).get().getState() == RequestState.FINISHED);
@@ -549,7 +553,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     schedule = "*/1 * * * * ?";
     requestResource.postRequest(request.toBuilder().setQuartzSchedule(Optional.of(schedule)).build());
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     Assert.assertTrue(!requestResource.getActiveRequests(false).isEmpty());
     Assert.assertTrue(requestManager.getRequest(requestId).get().getState() == RequestState.ACTIVE);
@@ -640,7 +644,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     SingularityRunNowRequest runNowRequest = new SingularityRunNowRequest(Optional.<String>absent(), Optional.<Boolean>absent(), Optional.<String>absent(), Optional.<List<String>>absent(), Optional.of(new Resources(2, 2, 0)), Optional.<Long>absent());
     requestResource.scheduleImmediately(requestId, runNowRequest);
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     SingularityPendingTask pendingTaskWithResourcs = taskManager.getPendingTasks().get(0);
     Assert.assertTrue(pendingTaskWithResourcs.getResources().isPresent());
@@ -660,7 +664,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     deployResource.deploy(new SingularityDeployRequest(new SingularityDeployBuilder(requestId, "d1").setCommand(Optional.of("cmd")).build(), Optional.<Boolean> absent(), Optional.<String> absent()));
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     deployChecker.checkDeploys();
 
@@ -679,7 +683,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     deployResource.deploy(new SingularityDeployRequest(new SingularityDeployBuilder(requestId, "d2").setCommand(Optional.of("cmd")).build(), Optional.<Boolean>absent(), Optional.<String> absent()));
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     deployChecker.checkDeploys();
 
@@ -712,7 +716,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     deployChecker.checkDeploys();
     Assert.assertEquals(2, requestManager.getSizeOfPendingQueue());
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     resourceOffers();
     Assert.assertEquals(2, taskManager.getActiveTaskIds().size());
@@ -726,7 +730,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     deployResource.deploy(new SingularityDeployRequest(new SingularityDeployBuilder(requestId, "d1").setCommand(Optional.of("cmd")).build(), Optional.<Boolean> absent(), Optional.<String> absent()));
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     deployChecker.checkDeploys();
 
@@ -739,7 +743,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     deployResource.deploy(new SingularityDeployRequest(new SingularityDeployBuilder(requestId, "d1").setCommand(Optional.of("cmd")).build(), Optional.<Boolean> absent(), Optional.<String> absent()));
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     deployChecker.checkDeploys();
     resourceOffers();
 
@@ -747,7 +751,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     slaveResource.decommissionSlave(taskManager.getActiveTasks().get(0).getSlaveId().getValue(), null);
 
-    scheduler.checkForDecomissions(stateCacheProvider.get());
+    scheduler.checkForDecomissions();
 
     cleaner.drainCleanupQueue();
     killKilledTasks();
@@ -777,7 +781,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     deployResource.deploy(new SingularityDeployRequest(new SingularityDeployBuilder(requestId, "d1").setCommand(Optional.of("cmd")).build(), Optional.<Boolean> absent(), Optional.<String> absent()));
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     deployChecker.checkDeploys();
     resourceOffers();
 
@@ -1024,8 +1028,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     initFirstDeploy();
 
     Assert.assertTrue(taskReconciliation.startReconciliation() == ReconciliationState.STARTED);
-    sleep(50);
-    Assert.assertTrue(!taskReconciliation.isReconciliationRunning());
+    Awaitility.await().atMost(10, TimeUnit.SECONDS).until(() -> !taskReconciliation.isReconciliationRunning());
 
     SingularityTask taskOne = launchTask(request, firstDeploy, 1, TaskState.TASK_STARTING);
     SingularityTask taskTwo = launchTask(request, firstDeploy, 2, TaskState.TASK_RUNNING);
@@ -1035,19 +1038,15 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     Assert.assertTrue(taskReconciliation.startReconciliation() == ReconciliationState.STARTED);
     Assert.assertTrue(taskReconciliation.startReconciliation() == ReconciliationState.ALREADY_RUNNING);
 
-    sleep(50);
-    Assert.assertTrue(taskReconciliation.isReconciliationRunning());
+    Awaitility.await().atMost(10, TimeUnit.SECONDS).until(() -> taskReconciliation.isReconciliationRunning());
 
     saveLastActiveTaskStatus(taskOne, Optional.of(buildTaskStatus(taskOne)), +1000);
 
-    sleep(50);
-    Assert.assertTrue(taskReconciliation.isReconciliationRunning());
+    Awaitility.await().atMost(10, TimeUnit.SECONDS).until(() -> taskReconciliation.isReconciliationRunning());
 
     saveLastActiveTaskStatus(taskTwo, Optional.of(buildTaskStatus(taskTwo)), +1000);
 
-    sleep(50);
-
-    Assert.assertTrue(!taskReconciliation.isReconciliationRunning());
+    Awaitility.await().atMost(10, TimeUnit.SECONDS).until(() -> !taskReconciliation.isReconciliationRunning());
   }
 
 
@@ -1073,22 +1072,22 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     List<SingularityTaskRequest> sortedRequestsByPriority = taskPrioritizer.getSortedDueTasks(requestsByPriority);
 
-    Assert.assertTrue(sortedRequestsByPriority.get(0).getRequest().getId().equals(highPriorityRequest.getId()));
-    Assert.assertTrue(sortedRequestsByPriority.get(1).getRequest().getId().equals(mediumPriorityRequest.getId()));
-    Assert.assertTrue(sortedRequestsByPriority.get(2).getRequest().getId().equals(lowPriorityRequest.getId()));
+    Assert.assertEquals(sortedRequestsByPriority.get(0).getRequest().getId(), highPriorityRequest.getId());
+    Assert.assertEquals(sortedRequestsByPriority.get(1).getRequest().getId(), mediumPriorityRequest.getId());
+    Assert.assertEquals(sortedRequestsByPriority.get(2).getRequest().getId(), lowPriorityRequest.getId());
 
     // A lower priority task that is long overdue should be run before a higher priority task
     now = System.currentTimeMillis();
     List<SingularityTaskRequest> requestsByOverdueAndPriority = Arrays.asList(
-      buildTaskRequest(lowPriorityRequest, lowPriorityDeploy, now - 60000), // 1 min overdue
-      buildTaskRequest(mediumPriorityRequest, mediumPriorityDeploy, now - 30000), // 30s overdue
+      buildTaskRequest(lowPriorityRequest, lowPriorityDeploy, now - 120000), // 2 min overdue
+      buildTaskRequest(mediumPriorityRequest, mediumPriorityDeploy, now - 30000), // 60s overdue
       buildTaskRequest(highPriorityRequest, highPriorityDeploy, now)); // Not overdue
 
     List<SingularityTaskRequest> sortedRequestsByOverdueAndPriority = taskPrioritizer.getSortedDueTasks(requestsByOverdueAndPriority);
 
-    Assert.assertTrue(sortedRequestsByOverdueAndPriority.get(0).getRequest().getId().equals(lowPriorityRequest.getId()));
-    Assert.assertTrue(sortedRequestsByOverdueAndPriority.get(1).getRequest().getId().equals(mediumPriorityRequest.getId()));
-    Assert.assertTrue(sortedRequestsByOverdueAndPriority.get(2).getRequest().getId().equals(highPriorityRequest.getId()));
+    Assert.assertEquals(sortedRequestsByOverdueAndPriority.get(0).getRequest().getId(), lowPriorityRequest.getId());
+    Assert.assertEquals(sortedRequestsByOverdueAndPriority.get(1).getRequest().getId(), mediumPriorityRequest.getId());
+    Assert.assertEquals(sortedRequestsByOverdueAndPriority.get(2).getRequest().getId(), highPriorityRequest.getId());
   }
 
   @Test
@@ -1278,7 +1277,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     cleaner.drainCleanupQueue();
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     resourceOffers();
 
     for (SingularityTaskId singularityTaskId : taskManager.getActiveTaskIds()) {
@@ -1316,7 +1315,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     // It acquires a lock on the bounce
     Assert.assertTrue("Lock on bounce should be acquired during bounce", requestManager.getExpiringBounce(requestId).isPresent());
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     resourceOffers();
 
     for (SingularityTaskId singularityTaskId : taskManager.getActiveTaskIds()) {
@@ -1523,11 +1522,11 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     requestResource.scheduleImmediately(requestId);
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     requestResource.scheduleImmediately(requestId);
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     Assert.assertEquals(2, taskManager.getPendingTaskIds().size());
 
@@ -1578,7 +1577,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
         )
     );
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     SingularityPendingTaskId task = taskManager.getPendingTaskIds().get(0);
     long runAt = task.getNextRunAt();
@@ -1605,7 +1604,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
         )
     );
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     SingularityPendingTaskId task = taskManager.getPendingTaskIds().get(0);
     long runAt = task.getNextRunAt();
@@ -1655,7 +1654,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     for (SingularityKilledTaskIdRecord taskId : taskManager.getKilledTaskIdRecords()) {
       Assert.assertTrue(taskId.getTaskId().getInstanceNo() > 2);
 
-      scheduler.drainPendingQueue(stateCacheProvider.get());
+      scheduler.drainPendingQueue();
     }
   }
 
@@ -1840,7 +1839,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     initFirstDeploy();
 
     requestResource.postRequest(request.toBuilder().setInstances(Optional.of(20)).build());
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     sms.resourceOffers(driver, Arrays.asList(createOffer(36, 12024)));
 
@@ -1859,12 +1858,12 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
   @Test
   public void testRequestedPorts() {
-    final SingularityDeployBuilder deployBuilder = dockerDeployWithPorts(3);
+    final SingularityDeployBuilder deployBuilder = dockerDeployWithPorts();
 
     initRequest();
-    initAndFinishDeploy(request, deployBuilder);
+    initAndFinishDeploy(request, deployBuilder, Optional.of(new Resources(1, 64, 3, 0)));
     requestResource.postRequest(request.toBuilder().setInstances(Optional.of(2)).build());
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     String[] portRangeWithNoRequestedPorts = {"65:70"};
     sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1", Optional.<String> absent(), Collections.<String, String>emptyMap(), portRangeWithNoRequestedPorts)));
@@ -1883,7 +1882,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     Assert.assertEquals(1, taskManager.getActiveTaskIds().size());
   }
 
-  private SingularityDeployBuilder dockerDeployWithPorts(int numPorts) {
+  private SingularityDeployBuilder dockerDeployWithPorts() {
     final SingularityDockerPortMapping literalMapping = new SingularityDockerPortMapping(Optional.<SingularityPortMappingType>absent(), 80, Optional.of(SingularityPortMappingType.LITERAL), 8080, Optional.<String>absent());
     final SingularityDockerPortMapping offerMapping = new SingularityDockerPortMapping(Optional.<SingularityPortMappingType>absent(), 81, Optional.of(SingularityPortMappingType.FROM_OFFER), 0, Optional.of("udp"));
     final SingularityContainerInfo containerInfo = new SingularityContainerInfo(
@@ -1898,7 +1897,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
           Optional.<Map<String, String>>of(ImmutableMap.of("env", "var=value")))
         ));
     final SingularityDeployBuilder deployBuilder = new SingularityDeployBuilder(requestId, "test-docker-ports-deploy");
-    deployBuilder.setContainerInfo(Optional.of(containerInfo)).setResources(Optional.of(new Resources(1, 64, numPorts, 0)));
+    deployBuilder.setContainerInfo(Optional.of(containerInfo));
     return deployBuilder;
   }
 
@@ -1915,7 +1914,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     requestManager.addToPendingQueue(new SingularityPendingRequest(requestId, "on_demand_deploy", System.currentTimeMillis(), Optional.<String>absent(), PendingType.ONEOFF,
       Optional.<List<String>>absent(), Optional.<String>absent(), Optional.<Boolean>absent(), Optional.<String>absent(), Optional.<String>absent()));
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     Assert.assertEquals(2, taskManager.getPendingTaskIds().size());
   }
@@ -1967,7 +1966,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     priorityResource.deleteActivePriorityFreeze();
 
     // launch task(s)
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     resourceOffers();
 
     // assert lowPriorityRequest has a new task running
@@ -1997,7 +1996,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     requestResource.scheduleImmediately(mediumPriorityRequest.getId());
 
     // drain pending queue
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     resourceOffers();
 
     // assert that lowPriorityRequest has a pending task
@@ -2012,7 +2011,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     Assert.assertEquals(SingularityDeleteResult.DELETED, priorityManager.deleteActivePriorityFreeze());
 
     // drain pending
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     resourceOffers();
 
     // check that both requests have active tasks
@@ -2029,7 +2028,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     requestManager.addToPendingQueue(new SingularityPendingRequest(requestId, firstDeployId, System.currentTimeMillis(), Optional.<String>absent(), PendingType.NEW_DEPLOY, Optional.<Boolean>absent(), Optional.<String>absent()));
 
     Assert.assertEquals(requestManager.getPendingRequests().size(), 1);
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     Assert.assertEquals(requestManager.getPendingRequests().size(), 0);
   }
@@ -2054,7 +2053,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     initAndFinishDeploy(request, "1");
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     final SingularityRequest newRequest = request.toBuilder()
       .setSchedule(Optional.of(newSchedule))
@@ -2067,7 +2066,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     deployChecker.checkDeploys();
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     Assert.assertEquals(newScheduleQuartz, requestManager.getRequest(requestId).get().getRequest().getQuartzScheduleSafe());
   }
@@ -2082,7 +2081,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     SingularityDeployRequest singularityDeployRequest = new SingularityDeployRequest(deploy, Optional.absent(), Optional.absent(), Optional.absent());
     deployResource.deploy(singularityDeployRequest);
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     SingularityPendingTask task1 = createAndSchedulePendingTask(firstDeployId);
 
@@ -2091,10 +2090,48 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     requestManager.addToPendingQueue(new SingularityPendingRequest(requestId, deploy.getId(), System.currentTimeMillis(), Optional.absent(), PendingType.IMMEDIATE,
         deploy.getSkipHealthchecksOnDeploy(), Optional.absent()));
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
     Assert.assertEquals(PendingType.IMMEDIATE, taskManager.getPendingTaskIds().get(0).getPendingType());
+  }
+
+  @Test
+  public void testSchedulerDropsMultipleScheduledTaskInstances() {
+    initScheduledRequest();
+
+    SingularityDeploy deploy = SingularityDeploy.newBuilder(requestId, firstDeployId)
+        .setCommand(Optional.of("sleep 100"))
+        .build();
+    SingularityDeployRequest singularityDeployRequest = new SingularityDeployRequest(deploy, Optional.absent(), Optional.absent(), Optional.absent());
+    deployResource.deploy(singularityDeployRequest);
+
+
+    scheduler.drainPendingQueue();
+    requestManager.addToPendingQueue(new SingularityPendingRequest(requestId, firstDeployId, Instant.now().plus(3, ChronoUnit.DAYS).toEpochMilli(), Optional.absent(), PendingType.NEW_DEPLOY, Optional.absent(), Optional.absent()));
+
+    SingularityRunNowRequest runNowRequest = new SingularityRunNowRequest(Optional.absent(), Optional.absent(), Optional.absent(), Optional.absent(), Optional.absent());
+    requestResource.scheduleImmediately(requestId, runNowRequest);
+
+
+
+
+    Assert.assertEquals("Both requests make it into the pending queue", 2, requestManager.getPendingRequests().size());
+    Assert.assertEquals(PendingType.IMMEDIATE, requestManager.getPendingRequests().get(0).getPendingType());
+    Assert.assertEquals(PendingType.NEW_DEPLOY, requestManager.getPendingRequests().get(1).getPendingType());
+
+
+    scheduler.drainPendingQueue();
+    Assertions.assertThat(taskManager.getPendingTaskIds())
+        .describedAs("Only the immediate request gets run")
+        .hasSize(1)
+        .extracting(SingularityPendingTaskId::getPendingType)
+        .containsExactly(PendingType.IMMEDIATE);
+
+
+    Assertions.assertThat(requestManager.getPendingRequests())
+        .describedAs("The scheduled request is dropped from the pending queue")
+        .hasSize(0);
   }
 
   @Test(expected = WebApplicationException.class)
@@ -2139,7 +2176,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     deployResource.deploy(new SingularityDeployRequest(deployGMT, Optional.<Boolean>absent(), Optional.<String>absent(), Optional.<SingularityRequest>absent()));
 
     deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     final long nextRunEST;
     final long nextRunGMT;
@@ -2243,7 +2280,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     cleaner.drainCleanupQueue();
     Assert.assertEquals(1, taskManager.getNumCleanupTasks());
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
     Assert.assertEquals(5, taskManager.getPendingTaskIds().size());
   }
 
@@ -2257,7 +2294,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     SingularityRunNowRequest runNowRequest = new SingularityRunNowRequest(Optional.<String>absent(), Optional.<Boolean>absent(), Optional.<String>absent(), Optional.<List<String>>absent(), Optional.of(new Resources(2, 2, 0)), Optional.<Long>absent());
     requestResource.scheduleImmediately(requestId, runNowRequest);
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     SingularityPendingTask pendingTaskWithResources = taskManager.getPendingTasks().get(0);
     Assert.assertTrue(pendingTaskWithResources.getResources().isPresent());
@@ -2283,7 +2320,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     SingularityRunNowRequest runNowRequest = new SingularityRunNowRequest(Optional.<String>absent(), Optional.<Boolean>absent(), Optional.<String>absent(), Optional.<List<String>>absent(), Optional.of(new Resources(2, 2, 0)), Optional.<Long>absent());
     requestResource.scheduleImmediately(requestId, runNowRequest);
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     SingularityPendingTask pendingTaskWithResources = taskManager.getPendingTasks().get(0);
     Assert.assertTrue(pendingTaskWithResources.getResources().isPresent());
@@ -2310,7 +2347,7 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     requestManager.addToPendingQueue(new SingularityPendingRequest(requestId, "on_demand_deploy", System.currentTimeMillis(), Optional.<String>absent(), PendingType.ONEOFF,
         Optional.<List<String>>absent(), Optional.<String>absent(), Optional.<Boolean>absent(), Optional.<String>absent(), Optional.<String>absent()));
 
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
 
     resourceOffers();
 
