@@ -12,6 +12,7 @@ import java.util.Set;
 import java.util.concurrent.CancellationException;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
+import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
 
 import org.apache.mesos.Protos.Attribute;
@@ -34,7 +35,6 @@ import org.junit.Before;
 import com.google.common.base.Optional;
 import com.google.common.base.Throwables;
 import com.google.inject.Inject;
-import com.google.inject.Provider;
 import com.google.inject.name.Named;
 import com.hubspot.baragon.models.BaragonRequestState;
 import com.hubspot.deploy.HealthcheckOptionsBuilder;
@@ -97,7 +97,7 @@ import com.ning.http.client.AsyncHttpClient;
 public class SingularitySchedulerTestBase extends SingularityCuratorTestBase {
 
   @Inject
-  protected Provider<SingularitySchedulerStateCache> stateCacheProvider;
+  protected SingularityLeaderCache leaderCache;
   @Inject
   protected SingularityMesosScheduler sms;
   @Inject
@@ -387,7 +387,7 @@ public class SingularitySchedulerTestBase extends SingularityCuratorTestBase {
   }
 
   protected void finishHealthchecks() {
-    for (Future<?> future : healthchecker.getHealthCheckFutures()) {
+    for (ScheduledFuture<?> future : healthchecker.getHealthCheckFutures()) {
       try {
         future.get();
       } catch (CancellationException ce) {
@@ -524,8 +524,12 @@ public class SingularitySchedulerTestBase extends SingularityCuratorTestBase {
     firstDeploy = initAndFinishDeploy(request, firstDeployId);
   }
 
+  protected void initFirstDeployWithResources(double cpus, double memoryMb) {
+    firstDeploy = initAndFinishDeployWithResources(request, firstDeployId, cpus, memoryMb);
+  }
+
   protected void initHCDeploy() {
-    firstDeploy = initAndFinishDeploy(request, new SingularityDeployBuilder(request.getId(), firstDeployId).setCommand(Optional.of("sleep 100")).setHealthcheck(Optional.of(new HealthcheckOptionsBuilder("http://uri").build())));
+    firstDeploy = initAndFinishDeploy(request, new SingularityDeployBuilder(request.getId(), firstDeployId).setCommand(Optional.of("sleep 100")).setHealthcheck(Optional.of(new HealthcheckOptionsBuilder("http://uri").build())), Optional.absent());
   }
 
   protected void initLoadBalancedDeploy() {
@@ -533,15 +537,21 @@ public class SingularitySchedulerTestBase extends SingularityCuratorTestBase {
         .setCommand(Optional.of("sleep 100"))
         .setServiceBasePath(Optional.of("/basepath"))
         .setLoadBalancerGroups(Optional.of(Collections.singleton("test")));
-    firstDeploy = initAndFinishDeploy(request, builder);
+    firstDeploy = initAndFinishDeploy(request, builder, Optional.absent());
   }
 
   protected SingularityDeploy initAndFinishDeploy(SingularityRequest request, String deployId) {
-    return initAndFinishDeploy(request, new SingularityDeployBuilder(request.getId(), deployId).setCommand(Optional.of("sleep 100")));
+    return initAndFinishDeploy(request, new SingularityDeployBuilder(request.getId(), deployId).setCommand(Optional.of("sleep 100")), Optional.absent());
   }
 
-  protected SingularityDeploy initAndFinishDeploy(SingularityRequest request, SingularityDeployBuilder builder) {
-    SingularityDeploy deploy = builder.build();
+  protected SingularityDeploy initAndFinishDeployWithResources(SingularityRequest request, String deployId, double cpus, double memoryMb) {
+    Resources r = new Resources(cpus, memoryMb, 0);
+
+    return initAndFinishDeploy(request, new SingularityDeployBuilder(request.getId(), deployId).setCommand(Optional.of("sleep 100")), Optional.of(r));
+  }
+
+  protected SingularityDeploy initAndFinishDeploy(SingularityRequest request, SingularityDeployBuilder builder, Optional<Resources> maybeResources) {
+    SingularityDeploy deploy = builder.setResources(maybeResources).build();
 
     SingularityDeployMarker marker = new SingularityDeployMarker(deploy.getRequestId(), deploy.getId(), System.currentTimeMillis(), Optional.<String> absent(), Optional.<String> absent());
 
@@ -674,7 +684,7 @@ public class SingularitySchedulerTestBase extends SingularityCuratorTestBase {
   protected void saveAndSchedule(SingularityRequestBuilder bldr) {
     requestManager.activate(bldr.build(), RequestHistoryType.UPDATED, System.currentTimeMillis(), Optional.<String> absent(), Optional.<String> absent());
     requestManager.addToPendingQueue(new SingularityPendingRequest(bldr.getId(), firstDeployId, System.currentTimeMillis(), Optional.<String> absent(), PendingType.UPDATED_REQUEST, Optional.<Boolean> absent(), Optional.<String> absent()));
-    scheduler.drainPendingQueue(stateCacheProvider.get());
+    scheduler.drainPendingQueue();
   }
 
   protected void saveLoadBalancerState(BaragonRequestState brs, SingularityTaskId taskId, LoadBalancerRequestType lbrt) {
