@@ -7,7 +7,7 @@ import {
   FetchRequestHistory
 } from '../../actions/api/history'
 
-import { DropdownButton, MenuItem } from 'react-bootstrap';
+import { DropdownButton, MenuItem, Glyphicon, Tooltip, OverlayTrigger } from 'react-bootstrap';
 import PauseButton from '../common/modalButtons/PauseButton';
 import UnpauseButton from '../common/modalButtons/UnpauseButton';
 import BounceButton from '../common/modalButtons/BounceButton';
@@ -21,8 +21,9 @@ class ActionDropdown extends React.Component {
     requests: PropTypes.object,
     fetchRequest: PropTypes.func.isRequired,
     fetchRequestHistory: PropTypes.func.isRequired,
-    fetchActiveTasksForRequest: PropTypes.func.isRequired
-  }
+    fetchActiveTasksForRequest: PropTypes.func.isRequired,
+    metadata: PropTypes.node
+  };
 
   constructor() {
     super();
@@ -54,41 +55,50 @@ class ActionDropdown extends React.Component {
   }
 
   render() {
-    const { group, requests } = this.props;
+    const { group, requests, metadata } = this.props;
+
+    const tooltip = (
+      <Tooltip id="request-group-tooltip">Apply to all requests in group</Tooltip>
+    );
 
     return (
-      <DropdownButton
-        bsStyle="primary"
-        title="Apply to all"
-        id="action-dropdown"
-        open={this.state.dropdownOpen}
-        onToggle={_.noop}
-        onClick={this.onMenuClick}
+      <OverlayTrigger placement="top" overlay={tooltip}>
+        <DropdownButton
+          bsStyle="default"
+          bsSize="small"
+          title={<Glyphicon glyph="option-vertical" />}
+          id="action-dropdown"
+          open={this.state.dropdownOpen}
+          onToggle={_.noop}
+          onClick={this.onMenuClick}
+          noCaret={true}
         >
+          <MenuItem header={true}>Request State</MenuItem>
+          <PauseButton requestId={group.requestIds} isScheduled={_.any(_.keys(requests), (requestId) => requests[requestId].requestType === 'SCHEDULED')} then={this.fetchRequestsAndHistoryAndActiveTasks}>
+            <MenuItem eventKey="1">Pause</MenuItem>
+          </PauseButton>
+          <UnpauseButton requestId={group.requestIds} then={this.fetchRequestsAndHistoryAndActiveTasks}>
+            <MenuItem eventKey="2">Unpause</MenuItem>
+          </UnpauseButton>
+          <MenuItem divider={true} />
 
-        <MenuItem header={true}>Request State</MenuItem>
-        <PauseButton requestId={group.requestIds} isScheduled={_.any(_.keys(requests), (requestId) => requests[requestId].requestType === 'SCHEDULED')} then={this.fetchRequestsAndHistoryAndActiveTasks}>
-          <MenuItem eventKey="1">Pause</MenuItem>
-        </PauseButton>
-        <UnpauseButton requestId={group.requestIds} then={this.fetchRequestsAndHistoryAndActiveTasks}>
-          <MenuItem eventKey="2">Unpause</MenuItem>
-        </UnpauseButton>
-        <MenuItem divider={true} />
+          <MenuItem header={true}>Healthchecks</MenuItem>
+          <EnableHealthchecksButton requestId={group.requestIds} then={this.fetchRequestsAndHistory}>
+            <MenuItem eventKey="3">Enable</MenuItem>
+          </EnableHealthchecksButton>
+          <DisableHealthchecksButton requestId={group.requestIds} then={this.fetchRequestsAndHistory}>
+            <MenuItem eventKey="4">Disable</MenuItem>
+          </DisableHealthchecksButton>
+          <MenuItem divider={true} />
 
-        <MenuItem header={true}>Healthchecks</MenuItem>
-        <EnableHealthchecksButton requestId={group.requestIds} then={this.fetchRequestsAndHistory}>
-          <MenuItem eventKey="3">Enable</MenuItem>
-        </EnableHealthchecksButton>
-        <DisableHealthchecksButton requestId={group.requestIds} then={this.fetchRequestsAndHistory}>
-          <MenuItem eventKey="4">Disable</MenuItem>
-        </DisableHealthchecksButton>
-        <MenuItem divider={true} />
+          <BounceButton requestId={group.requestIds} then={this.fetchRequestsAndHistoryAndActiveTasks}>
+            <MenuItem eventKey="5">Bounce</MenuItem>
+          </BounceButton>
 
-        <BounceButton requestId={group.requestIds} then={this.fetchRequestsAndHistoryAndActiveTasks}>
-          <MenuItem eventKey="5">Bounce</MenuItem>
-        </BounceButton>
-
-      </DropdownButton>
+          {metadata && <MenuItem divider={true} />}
+          {metadata}
+        </DropdownButton>
+      </OverlayTrigger>
     );
   }
 }
