@@ -1,15 +1,16 @@
 import React, { PropTypes } from 'react';
 import { connect } from 'react-redux';
-import rootComponent from '../../rootComponent';
-import { FetchSlaveUsages, FetchSlaves } from '../../actions/api/slaves';
-import { FetchSingularityStatus } from '../../actions/api/state';
-import { STAT_NAMES, WHOLE_NUMBER, HEALTH_SCALE_MAX } from './Constants';
-import Utils from '../../utils';
+import rootComponent from '../../../rootComponent';
+import { FetchSlaveUsages, FetchSlaves } from '../../../actions/api/slaves';
+import { FetchSingularityStatus } from '../../../actions/api/state';
+import { STAT_NAMES, WHOLE_NUMBER, HEALTH_SCALE_MAX } from '../Constants';
+import Utils from '../../../utils';
 import ResourceHealthData from './ResourceHealthData';
 import SlaveAggregates from './SlaveAggregates';
+import ClusterAggregates from './ClusterAggregates';
 
 const getSlaveInfo = (slaves, slaveUsage) => {
-  return _.findWhere(slaves, {'id' : slaveUsage.slaveId});
+  return _.findWhere(slaves, {'id': slaveUsage.slaveId});
 };
 
 const getUtilizationData = (slaves, slaveUsages) => {
@@ -26,7 +27,7 @@ const getUtilizationData = (slaves, slaveUsages) => {
   });
 };
 
-const SlaveUsage = ({slaves, slaveUsages, activeTasks}) => {
+const SlaveUsage = ({slaves, slaveUsages, activeTasks, clusterUtilization, totalRequests}) => {
   const activeSlaves = slaves.filter(Utils.isActiveSlave);
   const utilizationData = getUtilizationData(activeSlaves, slaveUsages);
 
@@ -42,12 +43,11 @@ const SlaveUsage = ({slaves, slaveUsages, activeTasks}) => {
     <div id="slave-usage-page">
       <h1>Slave Usage</h1>
       <div>
-        <SlaveAggregates slaves={activeSlaves} slaveUsages={slaveUsages} activeTasks={activeTasks} />
+        <SlaveAggregates slaves={activeSlaves} slaveUsages={slaveUsages} activeTasks={activeTasks} utilization={clusterUtilization} />
       </div>
-      <hr />
       <div id="slave-health">
         <h3>Slave health</h3>
-        <h4>Cpu</h4>
+        <h4>CPU</h4>
         <div className="cpu-health">
           {cpuHealthData}
         </div>
@@ -56,29 +56,37 @@ const SlaveUsage = ({slaves, slaveUsages, activeTasks}) => {
           {memoryHealthData}
         </div>
       </div>
+      <hr />
+      <h1>Cluster Utilization</h1>
+      <small className="last-updated">Last updated {Utils.timestampFromNow(clusterUtilization.timestamp)}</small>
+      <div>
+        <ClusterAggregates utilization={clusterUtilization} totalRequests={totalRequests} />
+      </div>
     </div>
   );
 };
 
 SlaveUsage.propTypes = {
-  slaveUsages : PropTypes.arrayOf(PropTypes.object),
-  slaves : PropTypes.arrayOf(PropTypes.object),
-  activeTasks : PropTypes.number
+  slaveUsages: PropTypes.arrayOf(PropTypes.object),
+  slaves: PropTypes.arrayOf(PropTypes.object),
+  activeTasks: PropTypes.number
 };
 
 function mapStateToProps(state) {
   return {
-    slaveUsages : state.api.slaveUsages.data,
-    slaves : state.api.slaves.data,
-    activeTasks : state.api.status.data.activeTasks
+    slaveUsages: state.api.slaveUsages.data,
+    slaves: state.api.slaves.data,
+    activeTasks: state.api.status.data.activeTasks,
+    clusterUtilization: state.api.utilization.data,
+    totalRequests: state.api.requests.data.length
   };
 }
 
 function mapDispatchToProps(dispatch) {
   return {
-    fetchSlaves : () => dispatch(FetchSlaves.trigger()),
-    fetchSlaveUsages : () => dispatch(FetchSlaveUsages.trigger()),
-    fetchSingularityStatus : () => dispatch(FetchSingularityStatus.trigger())
+    fetchSlaves: () => dispatch(FetchSlaves.trigger()),
+    fetchSlaveUsages: () => dispatch(FetchSlaveUsages.trigger()),
+    fetchSingularityStatus: () => dispatch(FetchSingularityStatus.trigger())
   };
 }
 
